Simplify array handling, make cell init more pythonic
This makes array indexes run from 0..self.cellHeight (width) as expected, removing a bunch of magic numbers.

It did require adding a new if test in "updateAllCells" "#skip indexes that are outside of array bounds", which makes
code slightly less efficient.  However, the code is much cleaner.

Not really doing anything special with border cells at the moment. They are envisioned as a way to help handle the
boundary cases where we could apply special rules near the edges to simulate an infinite array.